requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.0
frozenlist==1.4.1
grpclib==0.4.7
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
modal==0.61.108
multidict==6.0.5
Pillow==9.5.0
protobuf==4.25.3
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
rich==13.7.1
sigtools==4.0.1
sniffio==1.3.1
starlette==0.36.3
synchronicity==0.6.5
toml==0.10.2
typer==0.9.4
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing_extensions==4.10.0
watchfiles==0.21.0
yarl==1.9.4
